# specs/003-ai-chatbot-todo/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo AI Chatbot API
  description: |
    AI-powered conversational todo application API with natural language task management.

    Phase III adds conversational interface using OpenAI Agents SDK and MCP (Model Context Protocol).
    Users can manage tasks through natural language conversations.

    **Key Features:**
    - Natural language task management
    - Conversational AI interface
    - MCP-based tool architecture
    - Stateless server design
    - Conversation persistence
  version: 2.0.0
  contact:
    name: Todo Application Support
    email: support@todoapp.com
servers:
  - url: http://localhost:8001
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
...
  /api/{user_id}/chat:
    post:
      tags:
        - Chat
      summary: Send a message to the AI chatbot
      description: |
        Send a natural language message to the AI chatbot for task management.

        The chatbot can:
        - Create tasks: "Add a task to buy groceries"
        - List tasks: "Show me all my tasks" or "What's pending?"
        - Complete tasks: "Mark task 3 as complete"
        - Update tasks: "Change task 1 to 'Call mom tonight'"
        - Delete tasks: "Delete task 4"

        The endpoint is stateless - conversation history is loaded from the database
        for each request and the server maintains no in-memory session state.
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              new_conversation:
                summary: Start a new conversation
                value:
                  message: "Add a task to buy groceries"
              existing_conversation:
                summary: Continue existing conversation
                value:
                  conversation_id: 123
                  message: "What are my pending tasks?"
      responses:
        '200':
          description: Message processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                task_created:
                  summary: Task created via chat
                  value:
                    conversation_id: 123
                    response: "I've added 'Buy groceries' to your task list. Task ID: 42"
                    tool_calls: ["add_task"]
                    timestamp: "2026-01-26T10:00:00Z"
                task_listed:
                  summary: Tasks listed via chat
                  value:
                    conversation_id: 123
                    response: "You have 3 pending tasks:\n1. Buy groceries\n2. Call mom\n3. Finish report"
                    tool_calls: ["list_tasks"]
                    timestamp: "2026-01-26T10:01:00Z"
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                empty_message:
                  summary: Empty message
                  value:
                    error: "Message is required"
                    code: "INVALID_REQUEST"
                message_too_long:
                  summary: Message exceeds length limit
                  value:
                    error: "Message must be 2000 characters or less"
                    code: "INVALID_REQUEST"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - user does not have access to this conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Rate limit exceeded. Please wait before sending more messages."
                code: "RATE_LIMIT_EXCEEDED"
        '500':
          description: Internal server error - AI processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "AI service temporarily unavailable. Please try again."
                code: "AI_SERVICE_ERROR"
      security:
        - bearerAuth: []
  /api/{user_id}/conversations:
    get:
      tags:
        - Conversations
      summary: List all conversations for the authenticated user
      description: Returns all conversations for the specified user, ordered by last update (newest first).
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: Maximum number of conversations to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of conversations to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Conversations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationResponse'
                  total:
                    type: integer
                    description: Total number of conversations
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
  /api/{user_id}/conversations/{conversation_id}:
    get:
      tags:
        - Conversations
      summary: Get conversation details with message history
      description: Returns a specific conversation with its complete message history.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          description: Maximum number of messages to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Conversation retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    delete:
      tags:
        - Conversations
      summary: Delete a conversation and all its messages
      description: Permanently deletes a conversation and all associated messages.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Conversation deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

  # ==========================================
  # UTILITY ENDPOINTS
  # ==========================================
...
components:
  schemas:
...
    ChatRequest:
      type: object
      properties:
        conversation_id:
          type: integer
          nullable: true
          description: |
            Existing conversation ID to continue.
            If not provided, a new conversation will be created.
          example: 123
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: User's natural language message
          example: "Add a task to buy groceries"
      required:
        - message
    ChatResponse:
      type: object
      properties:
        conversation_id:
          type: integer
          description: The conversation ID (new or existing)
          example: 123
        response:
          type: string
          description: AI assistant's natural language response
          example: "I've added 'Buy groceries' to your task list. Task ID: 42"
        tool_calls:
          type: array
          items:
            type: string
          description: List of MCP tools invoked during processing
          example: ["add_task"]
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
          example: 2026-01-26T10:00:00Z

    # ==========================================
    # CONVERSATION SCHEMAS (Phase III)
    # ==========================================
    ConversationResponse:
      type: object
      properties:
        id:
          type: integer
          example: 123
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        created_at:
          type: string
          format: date-time
          example: 2026-01-26T09:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-26T10:00:00Z
        message_count:
          type: integer
          description: Total number of messages in conversation
          example: 10
        last_message_preview:
          type: string
          description: Preview of the last message (first 100 chars)
          example: "I've added 'Buy groceries' to your task list..."
    MessageResponse:
      type: object
      properties:
        id:
          type: integer
          example: 456
        conversation_id:
          type: integer
          example: 123
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        role:
          type: string
          enum: [user, assistant]
          description: Who sent the message
          example: user
        content:
          type: string
          description: Message content
          example: "Add a task to buy groceries"
        created_at:
          type: string
          format: date-time
          example: 2026-01-26T10:00:00Z
    ConversationDetailResponse:
      type: object
      properties:
        id:
          type: integer
          example: 123
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        created_at:
          type: string
          format: date-time
          example: 2026-01-26T09:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-01-26T10:00:00Z
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageResponse'
          description: Message history (ordered by created_at ascending)

    # ==========================================
    # ERROR SCHEMAS
    # ==========================================
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid email or password
        code:
          type: string
          description: Machine-readable error code
          example: INVALID_CREDENTIALS
        details:
          type: object
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          example: 2026-01-26T10:00:00Z

  # ==========================================
  # REUSABLE RESPONSES
  # ==========================================
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - user does not have access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ==========================================
  # SECURITY SCHEMES
  # ==========================================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token required for authentication.

        Obtain token via /api/auth/signup or /api/auth/signin endpoints.

        Include in Authorization header: `Authorization: Bearer <token>`
security:
  - bearerAuth: []
...